Add tcpdump & use latest senic-hub on branch
Install tcpdump on the dev images.
The motivation is to snif what packets are reaching the hub
and use it for a basis of automating test and reproducing issues

Latest senic-hub on branch:

For unknown reasons, the name of the recipe was different
that the package it creates. This reslutet in the needs
for the PROVIDES and RPROVIDES workaround.

It started breaking once a change that fetches always the
latest commit on a branch was done.

Instead of debugging the cause, I resulted to a best practice
and renamed the recipe to correspond to the pachage name.