Added multiprocessing-logging
The starting recipe was done with devtool:

    # On the build server
    devtool create-workspace
    devtool add  https://github.com/jruere/multiprocessing-logging
    devtool
    devtool edit-recipe multiprocessing-logging
        # inherit setuptools -> inherit setuptools3
        # delete the RDEPENDS that were wrongly detected
    devtool build multiprocessing-logging
    #Prerequisit: ssh root@10.10.10.100
    devtool deploy-target multiprocessing-logging root@10.10.10.100

    #On the 10.10.10.100 device
    python3
    import multiprocessing_logging

    #On the build server. Finishing
    devtool finish multiprocessing-logging /home/alan/meta-myproject/recipes-devtools/python

After which it was manually renamed to python3-multiprocessing-logging.

That failed at the process of creating the complete rootfs because of
some conflicts in packages trying to place test files in the same place
(don't know the details)

Removing the test files from the setup in the multiprocessing-logging
(that's what the patch is for) fixed the issue.